test working for hard coded axios request, refactor to server
