+++ .eslintrc.js
@@ -27,5 +27,6 @@ module.exports = {
     "react/prop-types": 0,
     "no-unused-vars": 0,
     "react/no-array-index-key": 0,
+    "react/jsx-filename-extension": [1, { extensions: [".js", ".jsx"] }],
   },
 };
